Step one: Making codebase parse
When running composer install, hhvm-autoload make an autoloadmap.
This map can not be made if the code uses unparseable constructs.
Think of instanceof and alike.
This code parses and composer install can complete.
Sadly, the testsuite fails 20 times, and errors 6 times.
My next step will be making the tests pass on both hhvm 4.27 and travis.